Acquisition protocol optimization
Optimization files for the micro-CT "CT-TB"
Estimated MTF cutoff frequency at 50 and 20% of the MTF for f=0 as a function of the magnification factor and the x-ray soucre power, that determines the focal spot size.
Relative signal compared to an ideal, flat-energy response detector as a function of x-ray tube working voltage and x-ray beam filtration
Mean response and mean variance of the detector as a function of exposure. Pixel variance as a function of pixel mean shows a linear relationship up to the saturation point of the detector